fix bug related to round
somehow, in certain python versions, round is not outputing int type, add int(round()) to ensure the data type

+++ examples/mixed_variable_benchmarks.py
@@ -257,7 +257,7 @@ elif benchmark == 'XG-MNIST':
             'min_split_loss': xc[1],
             'subsample': xc[2],
             'reg_lambda': xc[3],
-            'max_depth': round(xint[0]),
+            'max_depth': int(round(xint[0])),
             'num_class': 10  # the number of classes that exist in this datset
         }
 
@@ -320,7 +320,7 @@ elif benchmark == 'NAS-CIFAR10':
                 space[h.name] = x[i]
                 i += 1
             elif i == nc:
-                space[h.name] = round(x[i])
+                space[h.name] = int(round(x[i]))
                 i += 1
             else:
                 if x[i] == 0.:
