+++ src/pwasopt/acquisition.py
@@ -460,7 +460,7 @@ class active_learn:
         # constraints for one-hot encoded xint
         for i in range(nint):
             prob += (plp.lpSum(
-                xint[h][0] for h in range(round(sum(int_interval[:i])), round(sum(int_interval[:i + 1])))) == 1)
+                xint[h][0] for h in range(int(round(sum(int_interval[:i]))), int(round(sum(int_interval[:i + 1]))))) == 1)
 
         # constraints satisfaction
         if isLin_eqConstrained:
@@ -1079,7 +1079,7 @@ class active_learn:
 
         # constraints for one-hot encoded xint
         for i in range(nint):
-            prob += (plp.lpSum(xint[h][0] for h in range(round(sum(int_interval[:i])), round(sum(int_interval[:i + 1])))) == 1)
+            prob += (plp.lpSum(xint[h][0] for h in range(int(round(sum(int_interval[:i]))), int(round(sum(int_interval[:i + 1]))))) == 1)
 
 
         # constraints satisfaction
@@ -1458,7 +1458,7 @@ class active_learn:
 
         # constraints for one-hot encoded xint
         for i in range(nint):
-            prob += (plp.lpSum(xint[h][0] for h in range(round(sum(int_interval[:i])), round(sum(int_interval[:i + 1])))) == 1)
+            prob += (plp.lpSum(xint[h][0] for h in range(int(round(sum(int_interval[:i]))), int(round(sum(int_interval[:i + 1]))))) == 1)
 
         # constraints satisfaction
         if isLin_eqConstrained:
